api-gateway/requirements.txt
Flask==2.3.3
requests==2.31.0
prometheus-flask-exporter==0.23.0
tenacity==8.2.3
PyJWT==2.8.0
Flask-Limiter==3.5.0
redis==5.0.1
